error handing working now
