+++ src/types/global.d.ts
@@ -1,0 +1,10 @@
+// src/types/global.d.ts
+
+interface ErrorWithStatus extends Error {
+	status?: number;
+	errors?: any;
+	keyValue?: any;
+	code?: number;
+	path?: string;
+	value?: string;
+}
