+++ tsconfig.json
@@ -5,8 +5,9 @@
 		"rootDir": "./src",
 		"outDir": "./dist",
 		"strict": true,
-		"esModuleInterop": true
+		"esModuleInterop": true,
+		"typeRoots": ["./node_modules/@types", "./src/types"]
 	},
-	"include": ["src/**/*.ts"],
+	"include": ["src/**/*.ts", "src/types/global.d.ts"],
 	"exclude": ["node_modules"]
 }
